refactor(web): VrmViewer の hook をファイル分割する
VrmViewer.tsx(280行) の内部 hook を components/avatar/hooks/ へ逐語移動。

- use-latest-ref.ts: 最新 callback を ref に載せ替えるパターンを useLatestRef に
  共通化（useVrmLoader/useVrmAnimation の onErrorRef/onLoadedRef 重複を解消）。
- use-auto-blink.ts / use-vrm-loader.ts / use-expression-sync.ts /
  use-vrm-animation.ts に各 hook を分離。
- VrmViewer.tsx は VrmScene と公開 VrmViewer のみ残す。

GLTFLoader 生成/dispose は VRM と animation で dispose ロジックが大きく異なる
ため共通化は見送り（YAGNI）。公開 export は VrmViewer のみ不変、effect 依存・
cleanup・ref タイミングを完全保存。

nr validate / nr test (2661 pass, 0 fail) green。
